Resolved conflicting values for set_case analysis
In test mode, set_case_analysis values are collected from RC database. However, it
might happen that the same signal has different values in different timing modes,
such as tcb_tc 1 in shift and 0 in capture. In that case the shift value is preferred for
test-mode setup over the capture one.