Added debug_level procedure
If redirecting to a file it would have been nice to use the debug level
to put more less information. Access to execution_trace(debug-level) was
a bit cumbersome, thus this procedure. For future itr would make sense
to use the display_message to write to files too. Have not found an easy
way yet.